# ROCKBUSTER VISUAL FINDINGS and SQL statistics
This project was designed to be a pitch for Rockbuster client - a worldwide movie rental company. It contains a visual report for business managers at Rockbuster  with the most important findings related to rental rate numbers, rental duration, etc. .Also, it shows the countries where Rockbuster has the highest number of customers. Finally, it also contains a Data Dictionary.