# SQL Queries Project
As a Data Analyst student, this was my last project portraying  SQL queries learned and used to present a visual report for the business managers at ROCKBUSTER, a worldwide video rental company.  The content of this repository portrays the most important finding related to rental rate numbers, rental duration. Also, it shows the countries where ROCKBUSTER  has the highest number of customers.  This was my the first step in your journey towards a career in full-stack Data Analyst. 
Basically, you a list of SQL queries, a DataDictionary, a ppt with Tableau Public visualizations from a beginner level advancing slowly. More code snippets to be added soon.